Fix boot issue when customizing UEFI keys
After upgrading ISAR, the commit ilbers/isar@f7bce40 of ISAR introduced
a mandatory fixed value for SOURCE_DATE_EPOCH (1709565251,
corresponding to 2024-03-04 12:14:11). This fixed value is problematic
because it results in an incorrect signing time, causing the device to
fail to boot when customizing UEFI keys.

This commit sets SOURCE_DATE_EPOCH to the timestamp of the latest V tag.
This approach ensures that the signing time is meaningful and aligns
with the source code state, making it more logical and reliable.

Signed-off-by: Li Hua Qian <huaqian.li@siemens.com>