linux-iot2050: Ensure efivarfs is mounted as read-write
The efivarfs filesystem was being mounted as read-only because systemd
attempted to mount it before the TEE-based EFI runtime variable services
(TEE_STMM_EFI) were registered. This caused the efivarfs to lack write
capabilities, which is incorrect behavior.

To resolve this issue, the TEE_STMM_EFI driver is changed from a loadable
module to a built-in kernel component. This ensures that the TEE-based EFI
runtime variable services are initialized earlier during the boot process,
allowing efivarfs to detect write support and be mounted as read-write.

Signed-off-by: Li Hua Qian <huaqian.li@siemens.com>